Functionality to get users to follow items: - Sellers automatically follow their listed items     (change here to also refresh page after     creating a new listing, though it needs to be     uncommented) - On listings, queries are made to determine if an item is being followed by the user - User can add or remove items from the watchlist here
Small issues remaining:
- Need to refresh the page after adding or
removing an item from the watchlist to take effect

+++ views/watchlist_funcs.php
@@ -1,41 +1,31 @@
  <?php
+  require_once("../database/setup.php");
 
   // Follow or unfollow an item ()
-  if (!isset($_POST['functionname']) || !isset($_POST['arguments'])) {
+  if (!isset($_POST['functionname']) || !isset($_POST['argument_1']) || !isset($_POST['argument_2']) || !isset($_POST['argument_3'])) {
     return;
   }
+  // Extract variables from the POST variables
+  $item_id = $_POST['argument_1'];
+  $user_id = $_POST['argument_2'];
+  $follower_rows = $_POST['argument_3'];
 
-  // Extract arguments from the POST variables:
-  $item_id = $_POST['arguments'];
-
-  if (!isset($_SESSION['userId'])) {
-    $error = 'You must be logged in to add an item to your watchlist.';
-    echo  "<p><span class='font-weight-bold'>Error: </span> $error</p>";
-  } else {
-    $user_id = intval($_SESSION['userId']);
-    if ($user_id == 0) {
-      $error = 'Invalid userId.';
-      echo  "<p><span class='font-weight-bold'>Error: </span> $error</p>";
+  // Use argument to update the database (add or remove an item to the watchlist), returning success (if successful)
+  if ($_POST['functionname'] == 'add_to_watchlist') {
+    if ($follower_rows < 1) {
+      $query = "INSERT INTO FollowedItems (userId, itemId) VALUES ('$user_id', '$item_id')";
+      $db->query($query);
     }
-  }
-  // Preparing to query the database 
-  // Question: do I need this? I'd like to understand why we're being explicit about the database we're connecting to
-  $db->query("USE auction_site");
-
-  // Update the database (add or remove an item to the watchlist) and return success/failure
-  if ($_POST['functionname'] == "add_to_watchlist") {
-    $query = "INSERT INTO FollowedItems ($user_id, $item_id)";
+    $res = "success";
+  } elseif ($_POST['functionname'] == 'remove_from_watchlist') {
+    $query = "DELETE FROM FollowedItems WHERE userId = '$user_id' AND itemId = '$item_id'";
     $db->query($query);
-    $res = "You have successfully been added to the watchlist";
-  } else if ($_POST['functionname'] == "remove_from_watchlist") {
-    $query = "DELETE FROM FollowedItems ($user_id, $item_id)";
-    $db->query($query);
-    $res = "You have successfully been removed from the watchlist";
+    $res = "success";
   }
-  // Note: Echoing from this PHP function will return the value as a string.
-  // If multiple echo's in this file exist, they will concatenate together,
-  // so be careful. You can also return JSON objects (in string form) using
-  // echo json_encode($res).
+  //  Note: Echoing from this PHP function will return the value as a string.
+  //  If multiple echo's in this file exist, they will concatenate together,
+  //  so be careful. You can also return JSON objects (in string form) using
+  //  echo json_encode($res).
   echo $res;
 
   ?>
